#42: Refactored to use /lib directory
