--- utils/find-variations/README.md
@@ -1,11 +1,0 @@
-# Find Variations
-This tool uses a number of techniques to find variations between Data Dictionary metadata and the given file. 
-
-To use, a Metadata Report is required. 
-
-Usage:
-
-```
-reso-certification-utils findVariations -p /path/to/metadata-report.json
-```
-
